added background color support
