--- markdown-color.yaml
@@ -1,1 +1,0 @@
-enabled: true
