+++ themed-markdown-color.yaml
@@ -1,0 +1,24 @@
+enabled: true
+theme-blue: '#0000ff'
+theme-green: '#00ff00'
+theme-red: '#ff0000'
+theme-yellow: '#ffff00'
+theme-purple: '#ff00ff'
+theme-cyan: '#00ffff'
+theme-white: '#ffffff'
+theme-black: '#000000'
+theme-gray: '#808080'
+theme-orange: '#ffa500'
+theme-brown: '#a52a2a'
+theme-pink: '#ffc0cb'
+theme-maroon: '#800000'
+theme-olive: '#808000'
+theme-lime: '#00ff00'
+theme-teal: '#008080'
+theme-navy: '#000080'
+theme-silver: '#c0c0c0'
+theme-gold: '#ffd700'
+theme-violet: '#ee82ee'
+theme-magenta: '#ff00ff'
+theme-indigo: '#4b0082'
+theme-turquoise: '#40e0d0'
